+++ magnus/utils.py
@@ -9,7 +9,7 @@ from datetime import datetime
 from inspect import signature
 from pathlib import Path
 from string import Template as str_template
-from typing import Callable, Tuple, Union
+from typing import Callable, Tuple, Union, List, Mapping
 
 from ruamel.yaml import YAML  # type: ignore
 from stevedore import driver
@@ -269,7 +269,6 @@ def get_local_docker_image_id(image_name: str) -> str:
 
 
 def get_git_code_identity(run_log_store):
-    # TODO: Test this implementation when git is not present
     """
     Returns a code identity object for version controlled code.
 
@@ -387,17 +386,26 @@ def filter_arguments_for_func(func: Callable, parameters: dict, map_variable: di
         dict: The parameters matching the function signature
     """
     sign = signature(func)
-    filtered_parameters = {}
+    return filter_arguments_from_parameters(parameters=parameters,
+                                            signature_parameters=sign.parameters,
+                                            map_variable=map_variable)
+
+
+def filter_arguments_from_parameters(
+        parameters: dict, signature_parameters: Union[List, Mapping],
+        map_variable: dict = None) -> dict:
+    arguments = {}
+
     for param, value in parameters.items():
-        if param in sign.parameters:
-            filtered_parameters[param] = value
+        if param in signature_parameters:
+            arguments[param] = value
 
     if map_variable:
         for iterate_as, value in map_variable.items():
-            if iterate_as in sign.parameters:
-                filtered_parameters[iterate_as] = value
+            if iterate_as in signature_parameters:
+                arguments[iterate_as] = value
 
-    return filtered_parameters
+    return arguments
 
 
 def get_node_execution_command(executor, node, map_variable=None, over_write_run_id=None) -> str: